Refactor: 벨로그 제목 컴포넌트 분리
