+++ src/components/Button/Index.jsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
-import { primary1 } from '../../styles/color';
+import { buttonText, primary1, primary2 } from '../../styles/color';
 
-const Button = ({ text, color, background, hoverBackground, hoverColor, onClick }) => {
+const Button = ({ text, onClick, ...props }) => {
   return (
-    <Buttons onClick={onClick} text={text} color={color} background={background} hoverBackground={hoverBackground} hoverColor={hoverColor}>
+    <Buttons onClick={onClick} text={text} {...props}>
       {text}
     </Buttons>
   );
@@ -18,18 +18,18 @@ const Buttons = styled.button`
   height: 2rem;
   padding: 0px 1.25rem;
 
-  background: ${props => props.background};
+  background: ${primary2};
   outline: none;
   border: none;
   border-radius: 4px;
-  color: ${props => props.color};
+  color: ${buttonText};
   font-size: 1rem;
   font-weight: bold;
 
   &:hover {
     cursor: pointer;
-    background: ${props => props.hoverBackground};
-    color: ${props => props.hoverColor};
+    background: ${primary2};
+    color: ${buttonText};
   }
 `;
