+++ src/components/Setting/UserContents/Index.jsx
@@ -1,76 +1,29 @@
 import { useState, useEffect } from 'react';
 import Button from '../../Button';
 import EditButton from '../../EditButton';
+import Title from './Title';
 import SocialInfo from './SocialInfo';
 import Email from './Email';
 import EmailReceiveSetting from './EmailReceiveSetting';
 import Withdrawal from './Withdrawal';
 import styled from 'styled-components';
-import { backgroundElement1, border1, border3, border4, text2, text3, destructive1, destructive2, buttonText, primary1, primary2 } from '../../../styles/color';
+import { backgroundElement1, border1, border3, border4, text2, text3, buttonText, primary1, primary2 } from '../../../styles/color';
 
 const UserContents = () => {
-  const [title, setTitle] = useState('');
-  // const [social, setSocial] = useState('');
-  const [isModifyTitle, setIsModifyTitle] = useState(false);
-  // const [isModifyContents, setIsModifyContents] = useState(false);
-
-  const onModifyTitle = () => {
-    isModifyTitle ? setIsModifyTitle(false) : setIsModifyTitle(true);
-  };
-
-  // const onModifyContents = () => {
-  //   isModifyContents ? setIsModifyContents(false) : setIsModifyContents(true);
-  // };
-
-  useEffect(() => {
-    const contents = {
-      title: 'my.log',
-      social: { email: 'you8inpark@gmail.com', github: 'daydreamplace', twitter: 'eden', facebook: 'eden', url: 'dev-eden.shop' },
-    };
-    setTitle(contents.title);
-    // setSocial(contents.social);
-  }, []);
-
-  const getTitle = e => {
-    setTitle(e.target.value);
-  };
-
-  // const getSocial = e => {
-  //   setSocial(e.target.value);
-  // };
-
   return (
     <UserContentsContainer>
-      <div className='bottom-block'>
+      <div className='contents-box'>
         <div className='wrapper'>
           <div className='title-wrapper'>
             <h3>벨로그 제목</h3>
           </div>
           <div className='interval'>
-            {isModifyTitle ? (
-              <>
-                <input className='modify-input modify-title' type='text' placeholder='벨로그 제목' onChange={getTitle} value={title} />
-                <Button
-                  onClick={() => {
-                    setIsModifyTitle(false);
-                  }}
-                  text='저장'
-                  className='confirm-button'
-                />
-              </>
-            ) : (
-              <>
-                <div className='contents'>{title}</div>
-                <div className='edit-wrapper'>
-                  <EditButton text='수정' onClick={onModifyTitle} onChange={getTitle} value={title} />
-                </div>
-              </>
-            )}
+            <Title />
           </div>
         </div>
         <div className='desc'>개인 페이지의 좌측 상단에 나타나는 페이지 제목입니다.</div>
       </div>
-      <div className='bottom-block border'>
+      <div className='contents-box border'>
         <div className='wrapper'>
           <div className='title-wrapper'>
             <h3>소셜 정보</h3>
@@ -81,7 +34,7 @@ const UserContents = () => {
         </div>
         <div className='desc'>포스트 및 블로그에서 보여지는 프로필에 공개되는 소셜 정보입니다.</div>
       </div>
-      <div className='bottom-block border'>
+      <div className='contents-box border'>
         <div className='wrapper'>
           <div className='title-wrapper'>
             <h3>이메일 주소</h3>
@@ -90,24 +43,20 @@ const UserContents = () => {
         </div>
         <div className='desc'>회원 인증 또는 시스템에서 발송하는 이메일을 수신하는 주소입니다.</div>
       </div>
-      <div className='bottom-block border'>
+      <div className='contents-box border'>
         <div className='wrapper'>
           <div className='title-wrapper'>
             <h3>이메일 수신 설정</h3>
           </div>
-          <div className='contents'>
-            <EmailReceiveSetting />
-          </div>
+          <EmailReceiveSetting />
         </div>
       </div>
-      <div className='bottom-block border'>
+      <div className='contents-box border'>
         <div className='wrapper'>
           <div className='title-wrapper'>
             <h3>회원 탈퇴</h3>
           </div>
-          <div className='contents'>
-            <Withdrawal />
-          </div>
+          <Withdrawal />
         </div>
         <div className='desc'>탈퇴 시 작성하신 포스트 및 댓글이 모두 삭제되며 복구되지 않습니다.</div>
       </div>
@@ -122,7 +71,11 @@ const UserContentsContainer = styled.section`
     margin-top: 0rem;
   }
 
-  .bottom-block {
+  .border {
+    border-top: 1px solid ${border4};
+  }
+
+  .contents-box {
     padding-top: 1rem;
     padding-bottom: 1rem;
 
@@ -153,73 +106,6 @@ const UserContentsContainer = styled.section`
         display: flex;
         -webkit-box-align: center;
         align-items: center;
-
-        .modify-input {
-          flex: 1 1 0%;
-          display: block;
-          padding: 0.5rem;
-          margin-right: 1rem;
-
-          border: 1px solid ${border3};
-          border-radius: 4px;
-          background: ${backgroundElement1};
-          color: ${text2};
-          font-size: 1rem;
-          line-height: 1rem;
-          outline: none;
-
-          :focus {
-            border: 1px solid ${border1};
-          }
-        }
-
-        .confirm-button {
-          background: ${primary1};
-          color: ${buttonText};
-
-          &:hover {
-            background: ${primary2};
-          }
-        }
-
-        .edit-wrapper {
-          display: flex;
-          -webkit-box-align: center;
-          align-items: center;
-          margin-left: 1rem;
-        }
-      }
-
-      .contents {
-        /* flex: 1 1 0%;
-        font-size: 1rem;
-        color: ${text2};
-        line-height: 1.5; */
-
-        ul {
-          list-style: none;
-          padding: 0px;
-          margin: 0px;
-
-          li {
-            display: flex;
-            -webkit-box-align: center;
-            align-items: center;
-            margin-bottom: 0.5rem;
-
-            .icon {
-              width: 1rem;
-              height: 1rem;
-              font-size: 1rem;
-              margin-right: 0.5rem;
-              flex-shrink: 0;
-            }
-
-            span {
-              font-size: 1rem;
-            }
-          }
-        }
       }
     }
 
@@ -228,10 +114,6 @@ const UserContentsContainer = styled.section`
       color: ${text3};
       font-size: 0.875rem;
     }
-  }
-
-  .border {
-    border-top: 1px solid ${border4};
   }
 `;
